+++ OzzCodeGen/CodeEngines/CsModelClass/ModelPropertySetting.cs
@@ -1,4 +1,6 @@
 ﻿
+using OzzCodeGen.Definitions;
+
 namespace OzzCodeGen.CodeEngines.CsModelClass
 {
     public class ModelPropertySetting : BaseModelClassPropertySetting
@@ -8,5 +10,26 @@ namespace OzzCodeGen.CodeEngines.CsModelClass
             var entitySetting = (ModelClassEntitySetting)EntitySetting;
             return entitySetting?.CodeEngine;
         }
+
+        public bool IsSearchParameter
+        {
+            get
+            {
+                if (!_useInSearch.HasValue)
+                {
+                    _useInSearch = IsForeignKey() ||
+                        (PropertyDefinition.IsTypeString() &&
+                        ((StringProperty)PropertyDefinition).MaxLength < 257);
+                }
+                return _useInSearch.Value;
+            }
+            set
+            {
+                _useInSearch = value;
+                RaisePropertyChanged(nameof(IsSearchParameter));
+            }
+        }
+        private bool? _useInSearch;
+
     }
 }